requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
et-xmlfile==1.1.0
jdcal==1.4.1
numpy==1.24.2
opencv-python==4.7.0.68
openpyxl==2.5.1
packaging==22.0
pandas==1.5.3
pdf2image==1.16.0
Pillow==9.3.0
PyMySQL==1.0.2
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
XlsxWriter==3.0.8
